// src/enum/area.enum.ts
export enum DirectionEnum {
  North = 'North',
  East = 'East',
  Middle = 'Middle',
  South = 'South',
  Island = 'Island'
}

export enum CityEnum {
  Taipei = 'Taipei',
  NewTaipei = 'NewTaipei',
  Keelung = 'Keelung',
  Taoyuan = 'Taoyuan',
  Hsinchu = 'Hsinchu',
  HsinchuCounty = 'HsinchuCounty',
  YilanCounty = 'YilanCounty',
  MiaoliCounty = 'MiaoliCounty',
  Taichung = 'Taichung',
  ChanghuaCounty = 'ChanghuaCounty',
  NantouCounty = 'NantouCounty',
  YunlinCounty = 'YunlinCounty',
  Chiayi = 'Chiayi',
  ChiayiCounty = 'ChiayiCounty',
  Tainan = 'Tainan',
  Kaohsiung = 'Kaohsiung',
  PingtungCounty = 'PingtungCounty',
  PenghuCounty = 'PenghuCounty',
  HualienCounty = 'HualienCounty',
  TaitungCounty = 'TaitungCounty',
  KinmenCounty = 'KinmenCounty',
  LienchiangCounty = 'LienchiangCounty',
}